% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_GULF_DW.R
\name{connect_to_GULF_DW}
\alias{connect_to_GULF_DW}
\title{Connect to Gulf Power Data Warehouse on AWS}
\usage{
connect_to_GULF_DW(user, passwd)
}
\arguments{
\item{user}{A Character String, The User ID/SLID that needed to connect to EDP.}

\item{passwd}{A Character String, User's Password needed to Authenticate and
connect to the Gulf Power Data Warehouse.}
}
\value{
Returns A connection string of user ID and password to AWS which allows a user access to the Gulf Power Data Warehouse.
 Must have packages DBI and odbc installed in your Library.
}
\description{
Connect to Gulf Power Data Warehouse on AWS
}
\examples{
con1 <- connect_to_GULF_DW("ham0hv9", "User@2023")
}
